앵커 리팩토링 (#1533)
- 공통 로직 lib/anchor로 옮김
- Anchor 포지션 등 관리하는 로직은 Editor에서 Anchors 컴포넌트로 분리
