+++ apps/website/src/routes/website/_webview/editor/+page.svelte
@@ -21,7 +21,7 @@
   import { css, cx } from '$styled-system/css';
   import { center, flex } from '$styled-system/patterns';
   import { token } from '$styled-system/tokens';
-  import Anchor from './Anchor.svelte';
+  import Anchors from './Anchors.svelte';
   import { handleCaretMovement } from './caret';
   import FindReplace from './FindReplace.svelte';
   import Highlight from './Highlight.svelte';
@@ -782,7 +782,7 @@
         <Limit {$query} {editor} />
         <Spellcheck {editor} />
         <FindReplace {editor} />
-        <Anchor {doc} {editor} />
+        <Anchors {doc} {editor} />
       {/if}
     </div>
   </div>
